changed strategy for cart storage
Instead of saving all product information that I want to show on `Cart.vue` in localStorage. I opted back to only save the product id and the amount wanted in localStorage. This is because i still would of have needed to fetch all products to see how many was left in stock. The alternative would have been to save the stock amount in localStorage as well but that could lead to complications since the amount in stock is forever changing.